add: display items from parsed request in recycler view

+++ app/src/main/java/fr/isen/david/themaquereau/model/domain/Price.kt
@@ -6,10 +6,10 @@ data class Price(
     val id: Long,
     val id_pizza: Long,
     val id_size: Long,
-    val price: Int,
+    val price: Double,
     val create_date: Date,
     val update_date: Date,
     val size: String
 ) {
-    constructor() : this(0, 0, 0, 0, Date(), Date(), "")
+    constructor() : this(0, 0, 0, 0.0, Date(), Date(), "")
 }
